feat: add translateIfExists plugin, select i18n in a separate file
