+++ packages/docs/src/main.ts
@@ -1,24 +1,16 @@
 import { createApp } from 'vue'
 import App from './App.vue'
 import { router } from './router'
-import { createI18n } from 'vue-i18n'
-import { messages, DEFAULT_LANGUAGE } from './locales'
+import { DEFAULT_LANGUAGE } from './locales'
+import { i18n } from './locales/i18n'
 
 // plugin to change algolia colors according docs theme
 import AlgoliaColorPlugin from './components/sidebar/algolia-search/algolia-color-plugin'
 import { VuesticPlugin } from 'vuestic-ui/src/main'
 import { VuesticConfig } from './config/vuestic-config'
 import { useGtag } from './services/gtag'
+import { useTranslateIfExists } from './locales/translateIfExistsPlugin'
 import { useMeta } from './services/vue-meta'
-
-export const i18n = createI18n({
-  locale: DEFAULT_LANGUAGE,
-  fallbackLocale: 'en',
-  messages,
-  missing (_, key) {
-    return key
-  },
-})
 
 const app = createApp(App)
 
@@ -29,6 +21,7 @@ app.use(VuesticPlugin, VuesticConfig)
 
 useMeta(app)
 useGtag(app, router)
+useTranslateIfExists(app)
 
 app.mount('#app')
